+++ src/Sentry.Serilog/SentrySink.cs
@@ -161,8 +161,10 @@ internal sealed partial class SentrySink : ILogEventSink, IDisposable
                     ? exception?.Message ?? ""
                     : formatted,
                 context,
+                type: null,
                 data: data,
-                level: logEvent.Level.ToBreadcrumbLevel());
+                level: logEvent.Level.ToBreadcrumbLevel(),
+                hint: exception.ToHint());
         }
 
         // Read the options from the Hub, rather than the Sink's Serilog-Options, because 'EnableLogs' is declared in the base 'SentryOptions', rather than the derived 'SentrySerilogOptions'.