+++ src/Sentry/IHub.cs
@@ -23,7 +23,6 @@ public interface IHub : ISentryClient, ISentryScopeManager
     /// <remarks>
     /// Available options:
     /// <list type="bullet">
-    /// <item><see cref="Sentry.SentryOptions.EnableLogs"/></item>
     /// <item><see cref="Sentry.SentryOptions.SetBeforeSendLog(System.Func{SentryLog, SentryLog})"/></item>
     /// </list>
     /// </remarks>
@@ -35,7 +34,6 @@ public interface IHub : ISentryClient, ISentryScopeManager
     /// <remarks>
     /// Available options:
     /// <list type="bullet">
-    /// <item><see cref="Sentry.SentryOptions.EnableMetrics"/></item>
     /// <item><see cref="Sentry.SentryOptions.SetBeforeSendMetric(System.Func{SentryMetric, SentryMetric})"/></item>
     /// </list>
     /// </remarks>