+++ src/Sentry/Internal/Http/HttpTransport.cs
@@ -35,7 +35,8 @@ internal class HttpTransport : HttpTransportBase, ITransport
     /// </remarks>
     public virtual async Task SendEnvelopeAsync(Envelope envelope, CancellationToken cancellationToken = default)
     {
-        using var processedEnvelope = ProcessEnvelope(envelope);
+        // Caller to dispose
+        var processedEnvelope = ProcessEnvelope(envelope);
         if (processedEnvelope.Items.Count > 0)
         {
             using var request = CreateRequest(processedEnvelope);