+++ test/Sentry.Serilog.Tests/SentrySinkTests.cs
@@ -77,6 +77,29 @@ public partial class SentrySinkTests
     }
 
     [Fact]
+    public void EmitBreadcrumb_WithException_ProvidesExceptionInHint()
+    {
+        SentryHint hint = null;
+        _fixture.Scope.Options.SetBeforeBreadcrumb((breadcrumb, h) =>
+        {
+            hint = h;
+            return breadcrumb;
+        });
+        var expectedException = new Exception("expected message");
+
+        var sut = _fixture.GetSut();
+
+        // LogEventLevel.Warning is below the default MinimumEventLevel, so only a breadcrumb is added
+        var evt = new LogEvent(DateTimeOffset.UtcNow, LogEventLevel.Warning, expectedException,
+            MessageTemplate.Empty, Enumerable.Empty<LogEventProperty>());
+
+        sut.Emit(evt);
+
+        hint.Should().NotBeNull();
+        hint.Items[HintTypes.Exception].Should().BeSameAs(expectedException);
+    }
+
+    [Fact]
     public void Emit_SerilogSdk_Name()
     {
         var sut = _fixture.GetSut();